scripts/requirements.txt
# Dependencies for utility scripts
requests>=2.28.0
python-dotenv>=1.0.0
